Update test_harness clap args structure
Prepares it for adding a new subcommand
